# out-flask/openapi_server/models/footwear.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class Footwear(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    """
    allowed enum values
    """
    SHOES = "Shoes"
    SANDALS = "Sandals"
    HEELS = "Heels"
    FLIP_FLOPS = "Flip Flops"
    BALLERINAS = "Ballerinas"
    BOOTS = "Boots"
    SOCKS = "Socks"
    def __init__(self):  # noqa: E501
        """Footwear - a model defined in OpenAPI

        """
        self.openapi_types = {
        }

        self.attribute_map = {
        }

    @classmethod
    def from_dict(cls, dikt) -> 'Footwear':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Footwear of this Footwear.  # noqa: E501
        :rtype: Footwear
        """
        return util.deserialize_model(dikt, cls)